.github/workflows/requirements.txt
viam-sdk
asyncio
typesense